+++ ml_project/metrics/LogisticRegression_metrics.json
@@ -1,1 +1,1 @@
-{"accuracy_score": 0.8, "precision_score": 0.8214285714285714, "recall_score": 0.7666666666666667, "roc_auc_score": 0.8}
+{"accuracy_score": 0.75, "precision_score": 0.7142857142857143, "recall_score": 0.7407407407407407, "roc_auc_score": 0.7491582491582491}
