+++ ml_project/model_usage/data_manipulation/__init__.py
@@ -2,6 +2,7 @@ from .make_data import read_data, train_val_split
 from .process_features import (get_target, build_transformer,
                               build_categorical_transform_pipeline, build_num_transform_pipeline,
                               make_features)
+from .custom_transformer import RandomCustomTransformer
 
 __all__ = [
     "read_data",
@@ -11,4 +12,5 @@ __all__ = [
     "build_categorical_transform_pipeline",
     "build_num_transform_pipeline",
     "make_features",
+    "RandomCustomTransformer",
 ]
