+++ ml_project/tests/test_modules.py
@@ -18,31 +18,14 @@ import pandas as pd
 import pytest
 from pathlib import Path
 
-from model_usage.schemes import get_params_from_config
 from model_usage.data_manipulation import read_data, train_val_split, get_target, build_transformer, make_features
 from model_usage.model_training import train_model, create_fitted_pipeline, serialize_model
 from model_usage import train_pipeline, predict_pipeline
 from model_usage.model_testing import load_model, predict_labels, write_predictions
-from test_utils import generate_synthetic_data
+
+from fixtures import config_file, synthetic_data
 
 config_path = "configs/logreg_config.yaml"
-
-
-@pytest.fixture()
-def config_file():
-    model_config = get_params_from_config(config_path)
-    return model_config
-
-
-@pytest.fixture
-def synthetic_data(config_file):
-    gen_data_file = Path(config_file.paths.train_data_path.replace(".csv", "_synthetic.csv"))
-    if not gen_data_file.exists():
-        generate_synthetic_data(config_file.paths.train_data_path)
-    data = pd.read_csv(config_file.paths.train_data_path.replace(".csv", "_synthetic.csv"))
-    data[data[config_file.feature_params.target_col] >= 0.5] = 1
-    data[data[config_file.feature_params.target_col] < 0.5] = 0
-    return data
 
 
 def test_read_data(config_file, synthetic_data):
@@ -116,17 +99,3 @@ def test_write_predictions(config_file, synthetic_data):
     predicts = predict_labels(model, synthetic_data)
     write_predictions(predicts, config_file.paths.predictions_path)
     assert os.path.exists(config_file.paths.predictions_path)
-
-
-def test_end2end_training(config_file, synthetic_data):
-    train_pipeline(config_path)
-    assert os.path.exists(config_file.paths.model_path + f"/{config_file.train_params.model_type}_model.pkl")
-    assert os.path.exists(config_file.paths.metric_data_path + f"/{config_file.train_params.model_type}_metrics.json")
-    with open(config_file.paths.metric_data_path + f"/{config_file.train_params.model_type}_metrics.json", "r") as f:
-        metrics = json.load(f)
-    for metric in config_file.metrics:
-        assert metric in metrics
-
-
-def test_end2end_predicting():
-    predict_pipeline(config_path)
